+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ## 🛑 Breaking changes 🛑
 
 - Move configcheck.ValidateConfigFromFactories as internal function in service package (#3876).
+- Rename `configparser.Parser` as `config.Map` (#4075)
+
+## 💡 Enhancements 💡
+
 - Add Gen dependabot into CI (#4083)
 
 ## v0.36.0 Beta